refactor: 💡 Rename Logger.ts -> logger.ts
